Put DWARF functionality into a proper module
Continue what commit 9f3bd3f083a2 ("Create a proper src/gsym module
structure") and e13703754ac7 ("Properly organize ELF functionality")
started and restructure the DWARF code, putting it into a proper module
structure.
